+++ deltafi-ui/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- Metadata viewer table is now sortable
+
 ### Fixed
 - Bug on System Properties page causing name and description of PropertySets not to be displayed